requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
black==24.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.6
debugpy==1.6.6
decorator==5.1.1
distro==1.9.0
executing==1.2.0
fire==0.6.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipykernel==6.22.0
isort==5.13.2
jedi==0.18.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.1.0
jupyter_core==5.3.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langsmith==0.1.57
markdown-it-py==3.0.0
marshmallow==3.21.2
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
neo4j==5.20.0
numpy==1.26.4
openai==1.29.0
orjson==3.10.3
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.0.2
regex==2024.5.10
requests==2.31.0
rich==13.7.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
stack-data==0.6.2
tenacity==8.3.0
termcolor==2.4.0
tiktoken==0.7.0
tomli==2.0.1
tornado==6.2
tqdm==4.66.4
typing-inspect==0.9.0
urllib3==2.2.1
wcwidth==0.2.6
yarl==1.9.4
